[refactor] 코드 refactoring
