+++ index.html
@@ -8,7 +8,7 @@
       rel="stylesheet"
     />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Get to Know Wai</title>
+    <title>웨이 프로필</title>
   </head>
   <body>
     <div id="root"></div>
