+++ src/styles/aboutMe.styled.tsx
@@ -1,6 +1,8 @@
-import styled from 'styled-components';
+import styled, { keyframes } from 'styled-components';
 
-import { Container } from './container.styled';
+import { Container } from './common.styled';
+import { Text1, Text2 } from './common.styled';
+//AboutMe 세션 스타일
 export const AboutMeContainer = styled(Container)`
   height: 100%;
   display: flex;
@@ -20,10 +22,22 @@ export const ImageWrapper = styled.div`
   height: 100%;
 `;
 
-export const TextWrapper = styled.div`
-  flex: 2;
-  padding-left: 2rem;
-  @media (max-width: 768px) {
-    padding-left: 0;
+const zoomIn = keyframes`
+  0% {
+    transform: scale(0);
+    opacity: 0;
   }
+  100% {
+    transform: scale(1);
+    opacity: 1;
+  }
+`;
+export const AnimatedText1 = styled(Text1)`
+  animation: ${zoomIn} 2s ease forwards;
+  animation-delay: ${({ delay }) => delay ?? '0s'};
+`;
+
+export const AnimatedText2 = styled(Text2)`
+  animation: ${zoomIn} 2s ease forwards;
+  animation-delay: ${({ delay }) => delay ?? '0s'};
 `;
