+++ src/styles/header.styled.tsx
@@ -1,10 +1,9 @@
-// Styled-components for Header
 import styled from 'styled-components';
 
-import { Container } from './container.styled';
-import { Text2 } from './text.styled';
-import { theme } from './theme';
+import { Container } from './common.styled';
+import { Text2 } from './common.styled';
 
+//해더 Session 스타일
 export const HeaderContainer = styled(Container)`
   width: 100%;
   position: sticky;
@@ -13,7 +12,7 @@ export const HeaderContainer = styled(Container)`
   display: flex;
   justify-content: flex-end;
   align-items: center;
-  background-color: ${theme.color.primary};
+  background-color: ${({ theme }) => theme.color.primary};
 
   color: #000;
   height: 4rem;
@@ -34,8 +33,27 @@ export const NavItem = styled(Text2).attrs({ as: 'a' })`
     text-decoration: underline;
   }
 `;
+
+//연락처 Session 스타일
 export const ContactItem = styled(NavItem)`
   &:hover {
-    color: ${theme.color.bg};
+    color: ${({ theme }) => theme.color.bg};
+  }
+`;
+
+//Tmi Session 스타일
+export const SpinningImage = styled.img`
+  animation: spin 10s linear infinite;
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+
+  @keyframes spin {
+    0% {
+      transform: rotate(0deg);
+    }
+    100% {
+      transform: rotate(360deg);
+    }
   }
 `;
