+++ src/theme/globalStyle.styled.tsx
